+++ lib/public/Model.js
@@ -28,7 +28,7 @@ import { registerFrontLinkListener } from './components/common/navigation/frontL
 import { EosReportModel } from './views/EosReport/EosReportModel.js';
 import { AboutModel } from './views/About/About.js';
 import { StatisticsPageModel } from './views/Statistics/StatisticsPageModel.js';
-import { LhcPeriodsModel } from './views/lhcPeriods/overview/LhcPeriodsModel.js';
+import { LhcPeriodsModel } from './views/lhcPeriods/LhcPeriodsModel.js';
 
 /**
  * Root of model tree
