+++ lib/public/views/lhcPeriods/LhcPeriodsModel.js
@@ -17,7 +17,7 @@ import { LhcPeriodsOverviewModel } from './Overview/LhcPeriodsModel.js';
 /**
  * Model representing handlers for LhcPeriods
  */
-export default class LhcPeriodsModel extends Observable {
+export class LhcPeriodsModel extends Observable {
     /**
      * The constructor of the Overview model object
      */
