+++ lib/public/views/lhcPeriods/Overview/LhcPeriodsModel.js
