+++ lib/public/views/lhcPeriods/Overview/lhcPeriodsOverviewPanel.js
